Removing dependency on Gitlab instance
