+++ .gitlab-ci.yml
@@ -51,6 +51,10 @@ lint:docs:
   script: yarn lint:doc
 
 # Unit Tests
+test:unit:cli:
+  stage: test
+  script: yarn jest cli/test/unit  && yarn codecov -F cli
+
 test:unit:core:
   stage: test
   script: yarn jest core/test/unit && yarn codecov -F core
@@ -69,43 +73,39 @@ test:unit:utils:
   script: yarn jest utils/test/unit && yarn codecov -F utils
 
 # Integration Tests
-.test:integration: &integration
-  image:
-    name: docker/compose:latest
-    entrypoint: ['/bin/sh', '-c']
-  variables:
-    DOCKER_HOST: tcp://docker:2375
-    GITLAB_URL: http://docker:8080
-  services:
-    - docker:dind
-  stage: test
-  before_script:
-    # Install docker compose
-    - apk add --no-cache nodejs yarn git
+# .test:integration: &integration
+#   image:
+#     name: docker/compose:latest
+#     entrypoint: ['/bin/sh', '-c']
+#   variables:
+#     DOCKER_HOST: tcp://docker:2375
+#     GITLAB_URL: http://docker:8080
+#   services:
+#     - docker:dind
+#   stage: test
+#   before_script:
+#     # Install docker compose
+#     - apk add --no-cache nodejs yarn git
 
-    # Spin up container
-    - cd scripts
-    - docker-compose -f docker-compose.yml up -d
+#     # Spin up container
+#     - cd scripts
+#     - docker-compose -f docker-compose.yml up -d
 
-    # Verify Gitlab instance is up and running
-    - node probe.js
+#     # Verify Gitlab instance is up and running
+#     - node probe.js
 
-    # Get the personal token
-    - export PERSONAL_ACCESS_TOKEN=$(docker exec -i gitlab sh -c 'gitlab-rails r /mnt/init.rb')
+#     # Get the personal token
+#     - export PERSONAL_ACCESS_TOKEN=$(docker exec -i gitlab sh -c 'gitlab-rails r /mnt/init.rb')
 
-    - cd ..
-    - echo $PERSONAL_ACCESS_TOKEN
-    - echo $GITLAB_URL
-    - echo $CODECOV_TOKEN
+#     - cd ..
+#     - echo $PERSONAL_ACCESS_TOKEN
+#     - echo $GITLAB_URL
+#     - echo $CODECOV_TOKEN
 
 test:integration:browser:
   image: buildkite/puppeteer
   stage: test
   script: yarn jest browser/test/integration
-
-test:integration:cli:
-  <<: *integration
-  script: yarn jest cli/test/integration
 
 # test:integration:node:
 #   <<: *integration
