+++ packages/gitbeaker-browser/src/index.ts
@@ -2,7 +2,7 @@ import * as Gitbeaker from '@gitbeaker/core';
 import { modifyServices } from '@gitbeaker/requester-utils';
 import { Requester } from './KyRequester';
 
-const APIServices = modifyServices(Gitbeaker, [{ requester: Requester }]);
+const APIServices = modifyServices(Gitbeaker, { requester: Requester });
 
 export const {
   // Groups
