+++ packages/gitbeaker-requester-utils/src/RequesterUtils.ts
@@ -74,14 +74,16 @@ export function createInstance(optionsHandler, requestHandler): RequesterType {
   return requester;
 }
 
-export interface Constructable {
-  new (...args: any[]): any;
+export interface Constructable<T = {}> {
+  new (...args: any[]): T;
 }
 
 function extendClass<T extends Constructable>(Base: T, customConfig: object): T {
   return class extends Base {
     constructor(...options: any[]) {
-      super({ ...options, ...customConfig });
+      const [config, ...opts] = options;
+
+      super({ ...config, ...customConfig }, ...opts);
     }
   };
 }
